pyproject.toml
[project]
name = "autoflow"
version = "0.1.0"
description = "Desktop automation tool with screen recognition and keyboard/mouse control"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.6.0",
    "mss>=9.0.0",
    "opencv-python>=4.10.0",
    "pyautogui>=0.9.54",
    "numpy>=1.26.0",
    "aiosqlite>=0.22.1",
    "python-multipart>=0.0.22",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
ignore = ["B008"]  # Allow Depends() in function defaults (FastAPI pattern)
